Plugin rename for compatibility with wordpress store
