--- src/XUMM/Callback/AbstractHandler.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace Xrpl\XummForWoocommerce\XUMM\Callback;
-
-use Xrpl\XummSdkPhp\Response\GetPayload\XummPayload;
-use Xrpl\XummForWoocommerce\XUMM\Traits\XummPaymentGatewayTrait;
-use Xrpl\XummForWoocommerce\Woocommerce\XummPaymentGateway;
-
-/**
- * AbstractHandler Class
- *
- * The default behavior can be implemented inside a base handler class.
- *
- * @since      1.0.0
- * @package    Xrpl\XummForWoocommerce
- * @subpackage XUMM\Callback
- * @author     Andrei R <mdxico@gmail.com>
- */
-abstract class AbstractHandler implements Handler
-{
-    /** @use \Xrpl\XummForWoocommerce\XUMM\Traits\XummPaymentGatewayTrait */
-    use XummPaymentGatewayTrait;
-
-    /**
-     * @var XummPayload
-     */
-    protected XummPayload $payload;
-
-    /**
-     * Construct method
-     *
-     * @since     1.0.0
-     */
-    public function __construct(XummPaymentGateway $gateway, XummPayload $payload)
-    {
-        $this->setXummPaymentGateway($gateway);
-        $this->payload = $payload;
-    }
-
-    /**
-     * Default request handling
-     *
-     * @since     1.0.0
-     */
-    public function handle(): void
-    {
-        //
-    }
-}
