+++ tests/Helper/Mock/Responses/Transaction/USD.php
@@ -1,0 +1,256 @@
+<?php
+
+namespace Xrpl\XummForWoocommerce\Tests\Helper\Mock\Responses\Transaction;
+
+class USD
+{
+    public static function getResponse() : array
+    {
+        $object = new \stdClass();
+        $object->body = '{
+            "result":{
+               "Account":"rJR4MQt2egH9AmibZ8Hu5yTKVuLPv1xumm",
+               "Amount":{
+                  "currency":"USD",
+                  "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                  "value":"0.000001"
+               },
+               "Destination":"rJR4MQt2egH9AmibZ8Hu5yTKVuLPv1xumm",
+               "Fee":"15",
+               "Flags":2147483648,
+               "LastLedgerSequence":78012461,
+               "Memos":[
+                  {
+                     "Memo":{
+                        "MemoData":"4F726465722069643A203133362C207061696420776974682058554D4D",
+                        "MemoType":"68747470733A2F2F78756D6D2E6170702F"
+                     }
+                  }
+               ],
+               "Paths":[
+                  [
+                     {
+                        "currency":"USD",
+                        "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                        "type":48
+                     }
+                  ],
+                  [
+                     {
+                        "currency":"USD",
+                        "issuer":"rKiCet8SdvWxPXnAgYarFUXMh1zCPz432Y",
+                        "type":48
+                     },
+                     {
+                        "account":"rKiCet8SdvWxPXnAgYarFUXMh1zCPz432Y",
+                        "type":1
+                     },
+                     {
+                        "account":"rnAjjXdLxYZjmdYMhQq6eB2dKRfGZt4MxU",
+                        "type":1
+                     }
+                  ],
+                  [
+                     {
+                        "currency":"USD",
+                        "issuer":"rvYAfWj5gh67oV6fW32ZzP3Aw4Eubs59B",
+                        "type":48
+                     },
+                     {
+                        "account":"rvYAfWj5gh67oV6fW32ZzP3Aw4Eubs59B",
+                        "type":1
+                     },
+                     {
+                        "account":"rLhDJfjYWgGPtCetKkGHp4S2YBayDU81nV",
+                        "type":1
+                     }
+                  ],
+                  [
+                     {
+                        "currency":"USD",
+                        "issuer":"rvYAfWj5gh67oV6fW32ZzP3Aw4Eubs59B",
+                        "type":48
+                     },
+                     {
+                        "account":"rvYAfWj5gh67oV6fW32ZzP3Aw4Eubs59B",
+                        "type":1
+                     },
+                     {
+                        "account":"r4QdJy8GvanHmdX3TqWoGNyuGm1D5a2zow",
+                        "type":1
+                     }
+                  ]
+               ],
+               "SendMax":"3",
+               "Sequence":56678396,
+               "SigningPubKey":"0351C2E6AC9B0B3351C4B6406FEF627BF966FE009D33A9C7BAAB0FFEECC16E1656",
+               "TransactionType":"Payment",
+               "TxnSignature":"304402205656831320B6FE32C40375CE4AFF3910AA6E28A25D8155E951D506CE349304F302206E9FC70FA942FE4CBF6F5C80BCA3AA08EFC5D7FDFBA3B36EE58691DBFAC85852",
+               "date":730552422,
+               "hash":"876619C5488EE0335D1B50E0348EB4397E1B6954CAB5AE083381776D29929A4B",
+               "inLedger":78012453,
+               "ledger_index":78012453,
+               "meta":{
+                  "AffectedNodes":[
+                     {
+                        "ModifiedNode":{
+                           "FinalFields":{
+                              "Balance":{
+                                 "currency":"USD",
+                                 "issuer":"rrrrrrrrrrrrrrrrrrrrBZbvji",
+                                 "value":"-1.520615349126356"
+                              },
+                              "Flags":2228224,
+                              "HighLimit":{
+                                 "currency":"USD",
+                                 "issuer":"rJR4MQt2egH9AmibZ8Hu5yTKVuLPv1xumm",
+                                 "value":"100"
+                              },
+                              "HighNode":"0",
+                              "LowLimit":{
+                                 "currency":"USD",
+                                 "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                                 "value":"0"
+                              },
+                              "LowNode":"1c18"
+                           },
+                           "LedgerEntryType":"RippleState",
+                           "LedgerIndex":"668A09ACE1225DD7412BE032DA22A01C0248C81CDF2141155692132BE241FC85",
+                           "PreviousFields":{
+                              "Balance":{
+                                 "currency":"USD",
+                                 "issuer":"rrrrrrrrrrrrrrrrrrrrBZbvji",
+                                 "value":"-1.520614349126356"
+                              }
+                           },
+                           "PreviousTxnID":"E93F11C88FD3F7C117C781E0F8B5E17632564EA194F789FD5F4E80DD8D65C490",
+                           "PreviousTxnLgrSeq":73421947
+                        }
+                     },
+                     {
+                        "ModifiedNode":{
+                           "FinalFields":{
+                              "Account":"rnAHo4WfTcmbvwKPuKRyvAFPiDEznJJSxq",
+                              "Balance":"12710744918",
+                              "Flags":0,
+                              "MintedNFTokens":21,
+                              "OwnerCount":38,
+                              "Sequence":49819
+                           },
+                           "LedgerEntryType":"AccountRoot",
+                           "LedgerIndex":"66C063FF805A9EE6E7D4C7401170EA2EE1D4E0A7EFE85E260EDD5B05BC0891FA",
+                           "PreviousFields":{
+                              "Balance":"12710744915"
+                           },
+                           "PreviousTxnID":"603879774EC9CB1D6695AD3CCDD1A36B8A5E64D9E51506E50C6C7078BC86B1C9",
+                           "PreviousTxnLgrSeq":78012087
+                        }
+                     },
+                     {
+                        "ModifiedNode":{
+                           "FinalFields":{
+                              "Balance":{
+                                 "currency":"USD",
+                                 "issuer":"rrrrrrrrrrrrrrrrrrrrBZbvji",
+                                 "value":"-10015.33393457134"
+                              },
+                              "Flags":2228224,
+                              "HighLimit":{
+                                 "currency":"USD",
+                                 "issuer":"rnAHo4WfTcmbvwKPuKRyvAFPiDEznJJSxq",
+                                 "value":"1000000000"
+                              },
+                              "HighNode":"0",
+                              "LowLimit":{
+                                 "currency":"USD",
+                                 "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                                 "value":"0"
+                              },
+                              "LowNode":"4e3"
+                           },
+                           "LedgerEntryType":"RippleState",
+                           "LedgerIndex":"AE1D5C01A9398AC32DB7BAC44720D2C93AEF6FB64A561BDA897A01DBD4DE8BC6",
+                           "PreviousFields":{
+                              "Balance":{
+                                 "currency":"USD",
+                                 "issuer":"rrrrrrrrrrrrrrrrrrrrBZbvji",
+                                 "value":"-10015.33393557334"
+                              }
+                           },
+                           "PreviousTxnID":"EE4A9841F7466A95A360CE7534EE0A9F89F43EFE3D9D8951694F69887C2FCBB7",
+                           "PreviousTxnLgrSeq":78006406
+                        }
+                     },
+                     {
+                        "ModifiedNode":{
+                           "FinalFields":{
+                              "Account":"rnAHo4WfTcmbvwKPuKRyvAFPiDEznJJSxq",
+                              "BookDirectory":"F0B9A528CE25FE77C51C38040A7FEC016C2C841E74C1418D5B090A3C5EDF6DFC",
+                              "BookNode":"0",
+                              "Flags":0,
+                              "OwnerNode":"2",
+                              "Sequence":49813,
+                              "TakerGets":{
+                                 "currency":"USD",
+                                 "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                                 "value":"994.539998998"
+                              },
+                              "TakerPays":"2530636064"
+                           },
+                           "LedgerEntryType":"Offer",
+                           "LedgerIndex":"BCE6599E0485348D56D9D61EFA9B602DC0F4BB97FD1DF6C0C0C4EC2066ED354B",
+                           "PreviousFields":{
+                              "TakerGets":{
+                                 "currency":"USD",
+                                 "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                                 "value":"994.54"
+                              },
+                              "TakerPays":"2530636067"
+                           },
+                           "PreviousTxnID":"0534CE90908D6D4C9C80ADE97B2459A8E7A32064DA0F36E782AC0CC457281527",
+                           "PreviousTxnLgrSeq":78012073
+                        }
+                     },
+                     {
+                        "ModifiedNode":{
+                           "FinalFields":{
+                              "Account":"rJR4MQt2egH9AmibZ8Hu5yTKVuLPv1xumm",
+                              "Balance":"220685972",
+                              "Flags":0,
+                              "MessageKey":"020000000000000000000000005A46F885CC123730EC10E03F6E346927F9C5DD12",
+                              "OwnerCount":7,
+                              "Sequence":56678397
+                           },
+                           "LedgerEntryType":"AccountRoot",
+                           "LedgerIndex":"D4780D08AD9D89A0C0562A1B1D00FDE1886ACBCCBE91CEF4B6AFE73240345DB9",
+                           "PreviousFields":{
+                              "Balance":"220685990",
+                              "Sequence":56678396
+                           },
+                           "PreviousTxnID":"6AB5E5FA588B8A20AD78CCF23B9F752164AEB39D8A2C78C0C0B364222F9DBA50",
+                           "PreviousTxnLgrSeq":77822932
+                        }
+                     }
+                  ],
+                  "TransactionIndex":54,
+                  "TransactionResult":"tesSUCCESS",
+                  "delivered_amount":{
+                     "currency":"USD",
+                     "issuer":"rhub8VRN55s94qWKDv6jmDy1pUykJzF3wq",
+                     "value":"0.000001"
+                  }
+               },
+               "status":"success",
+               "validated":true,
+               "warnings":[
+                  {
+                     "id":1004,
+                     "message":"This is a reporting server.  The default behavior of a reporting server is to only return validated data. If you are looking for not yet validated data, include \"ledger_index : current\" in your request, which will cause this server to forward the request to a p2p node. If the forward is successful the response will include \"forwarded\" : \"true\""
+                  }
+               ]
+            }
+          }';
+
+        return json_decode(json_encode($object), true);
+    }
+}
